Move NetworkTab to gui.preftabs
